requirements.txt
# Glock - Model B Client Orchestrated Architecture
# Install with: pip install -r requirements.txt

# Core
pydantic>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0

# Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
litellm>=1.20.0

# Storage
redis>=5.0.0
asyncpg>=0.29.0

# Authentication & Security
pyjwt>=2.8.0
cryptography>=41.0.0

# CLI (optional - for TUI)
click>=8.1.0
rich>=13.0.0
websockets>=12.0
textual>=0.47.0
prompt-toolkit>=3.0.0
pygments>=2.17.0

# Web tools
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# PDF processing
pypdf>=3.0.0

# Image processing (optional)
pillow>=10.0.0

# Development
ruff>=0.1.0
mypy>=1.8.0
